Support compute quotas to optimize costs   - Compute quota API (renaming)
