--- api/src/main/java/com/epam/pipeline/mapper/compute/quota/ComputeQuotaRuleMapper.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2017-2026 EPAM Systems, Inc. (https://www.epam.com/)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.epam.pipeline.mapper.compute.quota;
-
-import com.epam.pipeline.dto.compute.quota.ComputeQuotaRule;
-import com.epam.pipeline.entity.compute.quota.ComputeQuotaRuleEntity;
-import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
-
-@Mapper(componentModel = "spring")
-public interface ComputeQuotaRuleMapper {
-
-    @Mapping(target = "action.type", source = "actionType")
-    @Mapping(target = "action.value", source = "actionValue")
-    @Mapping(target = "action.message", source = "actionMessage")
-    @Mapping(target = "action.perIncident", source = "perIncident")
-    ComputeQuotaRule toDto(ComputeQuotaRuleEntity entity);
-
-    @Mapping(target = "actionType", source = "action.type")
-    @Mapping(target = "actionValue", source = "action.value")
-    @Mapping(target = "actionMessage", source = "action.message")
-    @Mapping(target = "perIncident", source = "action.perIncident")
-    @Mapping(target = "createdDate", ignore = true)
-    @Mapping(target = "modifiedDate", ignore = true)
-    ComputeQuotaRuleEntity toEntity(ComputeQuotaRule rule);
-}
