+++ api/src/main/resources/messages.properties
@@ -794,13 +794,13 @@ error.aws.omics.doesnt.support.versioning=showVersion field is not empty, but AW
 error.aws.omics.store.incorrect.file.path=AWS Omics store supports: <file_id> or <file_id>/<source|source1|source2|index> as a path.
 
 #Compute quota rules
-error.compute.quota.rule.not.found=Compute quota rule with ID ''{0}'' was not found
-error.compute.quota.rule.name.empty=Compute quota rule name cannot be empty
-error.compute.quota.rule.filter.empty=Compute quota rule filter expression cannot be empty
-error.compute.quota.rule.action.empty=Compute quota rule action cannot be empty
-error.compute.quota.rule.unknown.field=Unknown filter field ''{0}''. Allowed fields: run.id, run.status, node.type, node.disk, docker.image, pipeline.id, pipeline.name, run.owner, run.spot, run.region_id, run.tag, run.owner.group
-error.compute.quota.rule.unsupported.operand=Operand ''{0}'' is not supported for field ''{1}''
-error.compute.quota.rule.duration.not.supported=Duration gate is not supported for field ''{0}''
+error.platform.usage.credits.rule.not.found=Compute quota rule with ID ''{0}'' was not found
+error.platform.usage.credits.rule.name.empty=Compute quota rule name cannot be empty
+error.platform.usage.credits.rule.filter.empty=Compute quota rule filter expression cannot be empty
+error.platform.usage.credits.rule.action.empty=Compute quota rule action cannot be empty
+error.platform.usage.credits.rule.unknown.field=Unknown filter field ''{0}''. Allowed fields: run.id, run.status, node.type, node.disk, docker.image, pipeline.id, pipeline.name, run.owner, run.spot, run.region_id, run.tag, run.owner.group
+error.platform.usage.credits.rule.unsupported.operand=Operand ''{0}'' is not supported for field ''{1}''
+error.platform.usage.credits.rule.duration.not.supported=Duration gate is not supported for field ''{0}''
 
 #Archive runs
 debug.archive.run.completed=Transferring runs to archive completed. Total archived runs count: {0}.
