+++ api/src/test/java/com/epam/pipeline/app/TestApplication.java
@@ -29,7 +29,7 @@ import com.epam.pipeline.manager.cloud.credentials.CloudProfileCredentialsManage
 import com.epam.pipeline.manager.cluster.InstanceOfferScheduler;
 import com.epam.pipeline.manager.cluster.performancemonitoring.ESMonitoringManager;
 import com.epam.pipeline.manager.cluster.pool.NodePoolUsageService;
-import com.epam.pipeline.manager.compute.quota.ComputeQuotaRuleService;
+import com.epam.pipeline.manager.credits.PlatformUsageCreditsRuleService;
 import com.epam.pipeline.manager.datastorage.lifecycle.DataStorageLifecycleManager;
 import com.epam.pipeline.manager.datastorage.providers.StorageEventCollector;
 import com.epam.pipeline.manager.ldap.LdapManager;
@@ -209,7 +209,7 @@ public class TestApplication {
     public GlobalSearchElasticHelper elasticHelper;
 
     @MockBean
-    public ComputeQuotaRuleService computeQuotaRuleService;
+    public PlatformUsageCreditsRuleService platformUsageCreditsRuleService;
 
     @Bean
     public EmbeddedServletContainerCustomizer containerCustomizer() throws FileNotFoundException {
