+++ api/src/test/java/com/epam/pipeline/manager/credits/PlatformUsageCreditsUpdateRuleServiceTest.java
@@ -1,0 +1,349 @@
+/*
+ * Copyright 2017-2026 EPAM Systems, Inc. (https://www.epam.com/)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.epam.pipeline.manager.credits;
+
+import com.epam.pipeline.common.MessageHelper;
+import com.epam.pipeline.controller.vo.FilterFieldVO;
+import com.epam.pipeline.dto.credits.*;
+import com.epam.pipeline.entity.credits.PlatformUsageCreditsUpdateRuleEntity;
+import com.epam.pipeline.mapper.credits.PlatformUsageCreditsRuleMapper;
+import com.epam.pipeline.repository.credits.PlatformUsageCreditsRuleRepository;
+import org.junit.Test;
+import org.mockito.ArgumentCaptor;
+
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.epam.pipeline.test.creator.credits.PlatformUsageCreditsRuleCreatorsUtils.ID;
+import static com.epam.pipeline.test.creator.credits.PlatformUsageCreditsRuleCreatorsUtils.platformUsageCreditsRule;
+import static com.epam.pipeline.test.creator.credits.PlatformUsageCreditsRuleCreatorsUtils.platformUsageCreditsRuleEntity;
+import static com.epam.pipeline.util.CustomAssertions.assertThrows;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.notNullValue;
+import static org.hamcrest.CoreMatchers.nullValue;
+import static org.junit.Assert.assertThat;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+
+public class PlatformUsageCreditsUpdateRuleServiceTest {
+
+    private final PlatformUsageCreditsRuleRepository repository = mock(PlatformUsageCreditsRuleRepository.class);
+    private final PlatformUsageCreditsRuleMapper mapper = mock(PlatformUsageCreditsRuleMapper.class);
+    private final MessageHelper messageHelper = mock(MessageHelper.class);
+    private final PlatformUsageCreditsRuleService manager =
+            new PlatformUsageCreditsRuleService(repository, mapper, messageHelper);
+
+    @Test
+    public void shouldLoadAll() {
+        final PlatformUsageCreditsUpdateRuleEntity entity = platformUsageCreditsRuleEntity();
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        doReturn(Collections.singletonList(entity)).when(repository).findAll();
+        doReturn(rule).when(mapper).toDto(entity);
+
+        final List<PlatformUsageCreditsUpdateRule> result = manager.loadAll();
+
+        assertThat(result.size(), is(1));
+        assertThat(result.get(0), is(rule));
+    }
+
+    @Test
+    public void shouldLoad() {
+        final PlatformUsageCreditsUpdateRuleEntity entity = platformUsageCreditsRuleEntity();
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        doReturn(entity).when(repository).findOne(ID);
+        doReturn(rule).when(mapper).toDto(entity);
+
+        assertThat(manager.load(ID), is(rule));
+    }
+
+    @Test
+    public void shouldFailLoadIfNotFound() {
+        doReturn(null).when(repository).findOne(ID);
+
+        assertThrows(IllegalArgumentException.class, () -> manager.load(ID));
+    }
+
+    @Test
+    public void shouldCreate() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        final PlatformUsageCreditsUpdateRuleEntity entity = platformUsageCreditsRuleEntity();
+        doReturn(entity).when(mapper).toEntity(rule);
+        doReturn(entity).when(repository).save(any(PlatformUsageCreditsUpdateRuleEntity.class));
+        doReturn(rule).when(mapper).toDto(entity);
+
+        final PlatformUsageCreditsUpdateRule result = manager.create(rule);
+
+        final ArgumentCaptor<PlatformUsageCreditsUpdateRuleEntity> captor =
+                ArgumentCaptor.forClass(PlatformUsageCreditsUpdateRuleEntity.class);
+        verify(repository).save(captor.capture());
+        assertThat(captor.getValue().getId(), nullValue());
+        assertThat(captor.getValue().getCreatedDate(), notNullValue());
+        assertThat(captor.getValue().getModifiedDate(), notNullValue());
+        assertThat(result, is(rule));
+    }
+
+    @Test
+    public void shouldSetDefaultStrategyTypeOnCreate() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        rule.setStrategyType(null);
+        final PlatformUsageCreditsUpdateRuleEntity entity = platformUsageCreditsRuleEntity();
+        entity.setStrategyType(null);
+        doReturn(entity).when(mapper).toEntity(rule);
+        doReturn(entity).when(repository).save(any(PlatformUsageCreditsUpdateRuleEntity.class));
+        doReturn(rule).when(mapper).toDto(entity);
+
+        manager.create(rule);
+
+        final ArgumentCaptor<PlatformUsageCreditsUpdateRuleEntity> captor =
+                ArgumentCaptor.forClass(PlatformUsageCreditsUpdateRuleEntity.class);
+        verify(repository).save(captor.capture());
+        assertThat(captor.getValue().getStrategyType(), is(PlatformUsageCreditsStrategyType.RUN_STATE));
+    }
+
+    @Test
+    public void shouldFailCreateIfNameIsBlank() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        rule.setName("");
+
+        assertThrows(IllegalArgumentException.class, () -> manager.create(rule));
+    }
+
+    @Test
+    public void shouldFailCreateIfFilterExpressionIsNull() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        rule.setFilterExpression(null);
+
+        assertThrows(IllegalArgumentException.class, () -> manager.create(rule));
+    }
+
+    @Test
+    public void shouldFailCreateIfActionIsNull() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        rule.setAction(null);
+
+        assertThrows(IllegalArgumentException.class, () -> manager.create(rule));
+    }
+
+    @Test
+    public void shouldUpdate() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        final PlatformUsageCreditsUpdateRuleEntity existing = platformUsageCreditsRuleEntity();
+        final PlatformUsageCreditsUpdateRuleEntity entity = platformUsageCreditsRuleEntity();
+        doReturn(existing).when(repository).findOne(ID);
+        doReturn(entity).when(mapper).toEntity(rule);
+        doReturn(entity).when(repository).save(any(PlatformUsageCreditsUpdateRuleEntity.class));
+        doReturn(rule).when(mapper).toDto(entity);
+
+        final PlatformUsageCreditsUpdateRule result = manager.update(ID, rule);
+
+        final ArgumentCaptor<PlatformUsageCreditsUpdateRuleEntity> captor =
+                ArgumentCaptor.forClass(PlatformUsageCreditsUpdateRuleEntity.class);
+        verify(repository).save(captor.capture());
+        assertThat(captor.getValue().getId(), is(ID));
+        assertThat(captor.getValue().getStrategyType(), is(existing.getStrategyType()));
+        assertThat(captor.getValue().getCreatedDate(), is(existing.getCreatedDate()));
+        assertThat(captor.getValue().getModifiedDate(), notNullValue());
+        assertThat(result, is(rule));
+    }
+
+    @Test
+    public void shouldFailUpdateIfNotFound() {
+        doReturn(null).when(repository).findOne(ID);
+
+        assertThrows(IllegalArgumentException.class, () -> manager.update(ID, platformUsageCreditsRule()));
+    }
+
+    @Test
+    public void shouldDelete() {
+        doReturn(true).when(repository).exists(ID);
+
+        manager.delete(ID);
+
+        verify(repository).delete(ID);
+    }
+
+    @Test
+    public void shouldFailDeleteIfNotFound() {
+        doReturn(false).when(repository).exists(ID);
+
+        assertThrows(IllegalStateException.class, () -> manager.delete(ID));
+    }
+
+    @Test
+    public void shouldGetKeywordsForAllRunFields() {
+        final List<FilterFieldVO> keywords = manager.getKeywords();
+
+        final List<String> expectedNames = Arrays.stream(PipelineRunField.values())
+                .flatMap(f -> f.getDisplayNames().stream())
+                .collect(Collectors.toList());
+        final List<String> actualNames = keywords.stream()
+                .map(FilterFieldVO::getFieldName)
+                .collect(Collectors.toList());
+        assertThat(actualNames, is(expectedNames));
+    }
+
+    @Test
+    public void shouldGetKeywordsWithSupportedOperands() {
+        final List<FilterFieldVO> keywords = manager.getKeywords();
+
+        keywords.forEach(kw -> assertThat(
+                "Expected non-empty operands for field: " + kw.getFieldName(),
+                kw.getSupportedOperands(), notNullValue()));
+        keywords.forEach(kw -> assertThat(
+                "Expected non-empty operands for field: " + kw.getFieldName(),
+                kw.getSupportedOperands().isEmpty(), is(false)));
+    }
+
+    // --- expression validation ---
+
+    @Test
+    public void shouldFailCreateIfFilterExpressionHasUnknownField() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        rule.setFilterExpression(leafExpression("unknown.field", "=", null));
+
+        assertThrows(IllegalArgumentException.class, () -> manager.create(rule));
+    }
+
+    @Test
+    public void shouldFailCreateIfExcludeExpressionHasUnknownField() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        rule.setExcludeExpression(leafExpression("unknown.field", "=", null));
+
+        assertThrows(IllegalArgumentException.class, () -> manager.create(rule));
+    }
+
+    @Test
+    public void shouldFailCreateIfOperandIsUnknownSymbol() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        // BOOLEAN field run.spot only supports = and !=; "??" is not a known operator at all
+        rule.setFilterExpression(leafExpression("run.spot", "??", null));
+
+        assertThrows(IllegalArgumentException.class, () -> manager.create(rule));
+    }
+
+    @Test
+    public void shouldFailCreateIfOperandIsNotSupportedForFieldType() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        // BOOLEAN field run.spot only supports = and !=; ">" is valid syntax but not for BOOLEAN
+        rule.setFilterExpression(leafExpression("run.spot", ">", null));
+
+        assertThrows(IllegalArgumentException.class, () -> manager.create(rule));
+    }
+
+    @Test
+    public void shouldFailCreateIfDurationSetOnNonDurationField() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        // run.status is ENUM and does NOT support duration
+        rule.setFilterExpression(leafExpression("run.status", "=", 24));
+
+        assertThrows(IllegalArgumentException.class, () -> manager.create(rule));
+    }
+
+    @Test
+    public void shouldCreateWithDurationOnTagField() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        // run.tag is the only field that supports duration
+        rule.setFilterExpression(leafExpression("run.tag", "=", 48));
+        final PlatformUsageCreditsUpdateRuleEntity entity = platformUsageCreditsRuleEntity();
+        doReturn(entity).when(mapper).toEntity(rule);
+        doReturn(entity).when(repository).save(any(PlatformUsageCreditsUpdateRuleEntity.class));
+        doReturn(rule).when(mapper).toDto(entity);
+
+        manager.create(rule);
+
+        verify(repository).save(any(PlatformUsageCreditsUpdateRuleEntity.class));
+    }
+
+    @Test
+    public void shouldValidateLeafNodesInsideAndExpression() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        final ConditionExpression unknown = leafExpression("unknown.field", "=", null);
+        final ConditionExpression valid = leafExpression("run.tag", "=", null);
+        final ConditionExpression and = new ConditionExpression();
+        and.setConditionType(ConditionType.AND);
+        and.setExpressions(Arrays.asList(valid, unknown));
+        rule.setFilterExpression(and);
+
+        assertThrows(IllegalArgumentException.class, () -> manager.create(rule));
+    }
+
+    // --- field normalization ---
+
+    @Test
+    public void shouldNormalizeFieldToLowerCaseOnCreate() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        rule.setFilterExpression(leafExpression("RUN.TAG", "=", 48));
+        final PlatformUsageCreditsUpdateRuleEntity entity = platformUsageCreditsRuleEntity();
+        doReturn(entity).when(mapper).toEntity(rule);
+        doReturn(entity).when(repository).save(any(PlatformUsageCreditsUpdateRuleEntity.class));
+        doReturn(rule).when(mapper).toDto(entity);
+
+        manager.create(rule);
+
+        assertThat(rule.getFilterExpression().getField(), is("run.tag"));
+    }
+
+    @Test
+    public void shouldNormalizeExcludeExpressionFieldToLowerCase() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        rule.setExcludeExpression(leafExpression("Node.Type", "=", null));
+        final PlatformUsageCreditsUpdateRuleEntity entity = platformUsageCreditsRuleEntity();
+        doReturn(entity).when(mapper).toEntity(rule);
+        doReturn(entity).when(repository).save(any(PlatformUsageCreditsUpdateRuleEntity.class));
+        doReturn(rule).when(mapper).toDto(entity);
+
+        manager.create(rule);
+
+        assertThat(rule.getExcludeExpression().getField(), is("node.type"));
+    }
+
+    @Test
+    public void shouldNormalizeFieldsInsideAndExpression() {
+        final PlatformUsageCreditsUpdateRule rule = platformUsageCreditsRule();
+        final ConditionExpression leaf1 = leafExpression("RUN.OWNER", "=", null);
+        final ConditionExpression leaf2 = leafExpression("RUN.SPOT", "=", null);
+        final ConditionExpression and = new ConditionExpression();
+        and.setConditionType(ConditionType.AND);
+        and.setExpressions(Arrays.asList(leaf1, leaf2));
+        rule.setFilterExpression(and);
+        final PlatformUsageCreditsUpdateRuleEntity entity = platformUsageCreditsRuleEntity();
+        doReturn(entity).when(mapper).toEntity(rule);
+        doReturn(entity).when(repository).save(any(PlatformUsageCreditsUpdateRuleEntity.class));
+        doReturn(rule).when(mapper).toDto(entity);
+
+        manager.create(rule);
+
+        assertThat(leaf1.getField(), is("run.owner"));
+        assertThat(leaf2.getField(), is("run.spot"));
+    }
+
+    private static ConditionExpression leafExpression(final String field,
+                                                      final String operand,
+                                                      final Integer duration) {
+        final ConditionExpression expr = new ConditionExpression();
+        expr.setConditionType(ConditionType.LOGICAL);
+        expr.setField(field);
+        expr.setOperand(operand);
+        expr.setValue("someValue");
+        expr.setDuration(duration);
+        return expr;
+    }
+}
