+++ api/src/test/java/com/epam/pipeline/test/acl/AclTestBeans.java
@@ -136,7 +136,7 @@ import com.epam.pipeline.manager.pipeline.runner.ConfigurationProviderManager;
 import com.epam.pipeline.manager.pipeline.runner.ConfigurationRunner;
 import com.epam.pipeline.manager.preference.PreferenceManager;
 import com.epam.pipeline.manager.preprocessing.NgsPreprocessingManager;
-import com.epam.pipeline.manager.compute.quota.ComputeQuotaRuleService;
+import com.epam.pipeline.manager.credits.PlatformUsageCreditsRuleService;
 import com.epam.pipeline.manager.quota.QuotaService;
 import com.epam.pipeline.manager.quota.RunLimitsService;
 import com.epam.pipeline.manager.region.CloudRegionManager;
@@ -580,7 +580,7 @@ public class AclTestBeans {
     protected QuotaService mockQuotaService;
 
     @MockBean
-    protected ComputeQuotaRuleService mockComputeQuotaRuleService;
+    protected PlatformUsageCreditsRuleService mockPlatformUsageCreditsRuleService;
 
     @MockBean
     protected UsersUsageReportService usersUsageReportService;
