+++ api/src/test/java/com/epam/pipeline/test/aspect/AspectTestBeans.java
@@ -71,7 +71,7 @@ import com.epam.pipeline.manager.billing.detail.EntityBillingDetailsLoader;
 import com.epam.pipeline.manager.cloud.CloudFacade;
 import com.epam.pipeline.manager.cluster.InstanceOfferScheduler;
 import com.epam.pipeline.manager.cluster.PodMonitor;
-import com.epam.pipeline.manager.compute.quota.ComputeQuotaRuleService;
+import com.epam.pipeline.manager.credits.PlatformUsageCreditsRuleService;
 import com.epam.pipeline.manager.contextual.handler.ContextualPreferenceHandler;
 import com.epam.pipeline.manager.datastorage.StorageQuotaTriggersManager;
 import com.epam.pipeline.manager.datastorage.lifecycle.DataStorageLifecycleManager;
@@ -552,5 +552,5 @@ public class AspectTestBeans {
     protected PipelineRunMetricsDao pipelineRunMetricsDao;
 
     @MockBean
-    protected ComputeQuotaRuleService computeQuotaRuleService;
+    protected PlatformUsageCreditsRuleService platformUsageCreditsRuleService;
 }
