+++ api/src/test/java/com/epam/pipeline/test/web/ControllerTestBeans.java
@@ -34,7 +34,7 @@ import com.epam.pipeline.acl.pipeline.issue.GitlabIssueApiService;
 import com.epam.pipeline.acl.plugin.PluginAssignmentService;
 import com.epam.pipeline.acl.plugin.PluginService;
 import com.epam.pipeline.acl.preprocessing.NgsPreprocessingApiService;
-import com.epam.pipeline.acl.compute.quota.ComputeQuotaRuleApiService;
+import com.epam.pipeline.acl.credits.PlatformUsageCreditsRuleApiService;
 import com.epam.pipeline.acl.quota.QuotaApiService;
 import com.epam.pipeline.acl.report.ReportApiService;
 import com.epam.pipeline.acl.resource.StaticResourceApiService;
@@ -277,7 +277,7 @@ public class ControllerTestBeans {
     protected QuotaApiService quotaApiService;
 
     @MockBean
-    protected ComputeQuotaRuleApiService computeQuotaRuleApiService;
+    protected PlatformUsageCreditsRuleApiService platformUsageCreditsRuleApiService;
 
     @MockBean
     protected ReportApiService reportApiService;
